+++ Core/ChatManager.lua
@@ -147,7 +147,8 @@ function QueueHandler:UpdateBandwidth()
 end
 
 function QueueHandler:DoesChatTypeRequireHardwareInput(chatType)
-    return Utils.IsInOpenWorld() and HARDWARE_RESTRICTED_CHAT_TYPES[chatType];
+    return (Utils.IsInOpenWorld() or C_Garrison.IsOnGarrisonMap())
+		and HARDWARE_RESTRICTED_CHAT_TYPES[chatType];
 end
 
 function QueueHandler:TrySendMessage(entry)