+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 ## Added
 - Messages being sent using TRP3's NPC speech now persist the NPC speech formatting through all messages.
 	- You can toggle this behavior in the Chattery settings (found in the Addon Compartment or through `/chattery`)
+
+## Fixed
+- Fixed an issue sending certain types of chat messages while in a garrison.