+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Messages being sent using TRP3's NPC speech now persist the NPC speech formatting through all messages.
 	- You can toggle this behavior in the Chattery settings (found in the Addon Compartment or through `/chattery`)
 - Added hints to settings in the settings UI
+- Added toggles for auto-capitalization and punctuation
+	- When enabled, Chattery will automatically capitalize the beginning of the first message, and add a period to the end of the last message.
 
 ## Fixed
 - Fixed an issue sending certain types of chat messages while in a garrison.