+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.7.9
+# 0.8.0
 
 ## Added
 - Messages being sent using TRP3's NPC speech now persist the NPC speech formatting through all messages.
@@ -6,6 +6,8 @@
 - Added hints to settings in the settings UI
 - Added toggles for auto-capitalization and punctuation
 	- When enabled, Chattery will automatically capitalize the beginning of the first message, and add a period to the end of the last message.
+- Added undo support
+	- If you accidentally delete your message or close the edit box, you can press Ctrl+Z to undo!
 
 ## Fixed
 - Fixed an issue sending certain types of chat messages while in a garrison.